testing out queries in dashboard
